# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cross-Border Sentinel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnalyzeRequest {
  /** News article text to analyze */
  text: string;
  /** Optional image to verify */
  image?: Blob;
}

export interface TextAnalysis {
  /** Text credibility score 0-100 */
  score: number;
  /** Model confidence 0-1 */
  confidence: number;
  /** Detected anomaly flags */
  flags: string[];
  /** Credibility indicators */
  positive_signals: string[];
  /** Whether strong credibility signals are present */
  has_strong_positive: boolean;
  /** Whether strong anomaly signals are present */
  has_strong_negative: boolean;
  /** Whether major anomalies (sensational/extraordinary) were detected */
  has_major_anomalies: boolean;
  /** Accumulated anomaly weight — Hard Negative Rule fires at ≥40 */
  anomaly_score: number;
  /** Count of unverified sourcing instances */
  negative_uncertainty_count: number;
  /** Count of responsible uncertainty phrases */
  positive_uncertainty_count: number;
}

export interface ImageAnalysis {
  /** Image credibility score 0-100 */
  score: number;
  has_image: boolean;
  flags: string[];
  positive_signals: string[];
}

export type AnalysisResultPrediction =
  (typeof AnalysisResultPrediction)[keyof typeof AnalysisResultPrediction];

export const AnalysisResultPrediction = {
  Real: "Real",
  Fake: "Fake",
  Uncertain: "Uncertain",
} as const;

export interface AnalysisResult {
  /** Record ID in history */
  id: number;
  /** Overall credibility score 0-100 */
  credibility_score: number;
  prediction: AnalysisResultPrediction;
  /** Human-readable explanation points */
  explanation: string[];
  text_analysis: TextAnalysis;
  image_analysis: ImageAnalysis;
  analyzed_at: string;
  /** First 200 chars of analyzed text */
  text_preview: string;
  /** Full source text (returned by detail endpoint) */
  text_content?: string;
}

export type HistoryItemPrediction =
  (typeof HistoryItemPrediction)[keyof typeof HistoryItemPrediction];

export const HistoryItemPrediction = {
  Real: "Real",
  Fake: "Fake",
  Uncertain: "Uncertain",
} as const;

export interface HistoryItem {
  id: number;
  credibility_score: number;
  prediction: HistoryItemPrediction;
  text_preview: string;
  analyzed_at: string;
  has_image: boolean;
}

export interface HistoryResponse {
  items: HistoryItem[];
  total: number;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}