# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cross-Border Sentinel API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: analysis
    description: Fake news analysis operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /analyze:
    post:
      operationId: analyzeContent
      tags: [analysis]
      summary: Analyze news content for credibility
      description: Analyzes text and optionally an image to produce a credibility score
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/AnalyzeRequest"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /history:
    get:
      operationId: getHistory
      tags: [analysis]
      summary: Get analysis history
      description: Returns a list of previously analyzed content
      responses:
        "200":
          description: List of analysis history records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"
    delete:
      operationId: clearHistory
      tags: [analysis]
      summary: Clear all analysis history
      responses:
        "200":
          description: History cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /history/{id}:
    get:
      operationId: getHistoryItem
      tags: [analysis]
      summary: Get full detail for a single history record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Full analysis detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteHistoryItem
      tags: [analysis]
      summary: Delete a specific history item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AnalyzeRequest:
      type: object
      properties:
        text:
          type: string
          description: News article text to analyze
        image:
          type: string
          format: binary
          description: Optional image to verify
      required:
        - text
    TextAnalysis:
      type: object
      properties:
        score:
          type: number
          description: Text credibility score 0-100
        confidence:
          type: number
          description: Model confidence 0-1
        flags:
          type: array
          items:
            type: string
          description: Detected anomaly flags
        positive_signals:
          type: array
          items:
            type: string
          description: Credibility indicators
        has_strong_positive:
          type: boolean
          description: Whether strong credibility signals are present
        has_strong_negative:
          type: boolean
          description: Whether strong anomaly signals are present
        has_major_anomalies:
          type: boolean
          description: Whether major anomalies (sensational/extraordinary) were detected
        anomaly_score:
          type: number
          description: Accumulated anomaly weight — Hard Negative Rule fires at ≥40
        negative_uncertainty_count:
          type: number
          description: Count of unverified sourcing instances
        positive_uncertainty_count:
          type: number
          description: Count of responsible uncertainty phrases
      required:
        - score
        - confidence
        - flags
        - positive_signals
        - has_strong_positive
        - has_strong_negative
        - has_major_anomalies
        - anomaly_score
        - negative_uncertainty_count
        - positive_uncertainty_count
    ImageAnalysis:
      type: object
      properties:
        score:
          type: number
          description: Image credibility score 0-100
        has_image:
          type: boolean
        flags:
          type: array
          items:
            type: string
        positive_signals:
          type: array
          items:
            type: string
      required:
        - score
        - has_image
        - flags
        - positive_signals
    AnalysisResult:
      type: object
      properties:
        id:
          type: integer
          description: Record ID in history
        credibility_score:
          type: number
          description: Overall credibility score 0-100
        prediction:
          type: string
          enum: [Real, Fake, Uncertain]
        explanation:
          type: array
          items:
            type: string
          description: Human-readable explanation points
        text_analysis:
          $ref: "#/components/schemas/TextAnalysis"
        image_analysis:
          $ref: "#/components/schemas/ImageAnalysis"
        analyzed_at:
          type: string
          format: date-time
        text_preview:
          type: string
          description: First 200 chars of analyzed text
        text_content:
          type: string
          description: Full source text (returned by detail endpoint)
      required:
        - id
        - credibility_score
        - prediction
        - explanation
        - text_analysis
        - image_analysis
        - analyzed_at
        - text_preview
    HistoryItem:
      type: object
      properties:
        id:
          type: integer
        credibility_score:
          type: number
        prediction:
          type: string
          enum: [Real, Fake, Uncertain]
        text_preview:
          type: string
        analyzed_at:
          type: string
          format: date-time
        has_image:
          type: boolean
      required:
        - id
        - credibility_score
        - prediction
        - text_preview
        - analyzed_at
        - has_image
    HistoryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/HistoryItem"
        total:
          type: integer
      required:
        - items
        - total
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message