# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cross-Border Sentinel API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResultPrediction } from "./analysisResultPrediction";
import type { ImageAnalysis } from "./imageAnalysis";
import type { TextAnalysis } from "./textAnalysis";

export interface AnalysisResult {
  /** Record ID in history */
  id: number;
  /** Overall credibility score 0-100 */
  credibility_score: number;
  prediction: AnalysisResultPrediction;
  /** Human-readable explanation points */
  explanation: string[];
  text_analysis: TextAnalysis;
  image_analysis: ImageAnalysis;
  analyzed_at: Date;
  /** First 200 chars of analyzed text */
  text_preview: string;
  /** Full source text (returned by detail endpoint) */
  text_content?: string;
}